added pagination

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const Product = require("../models/product"); 
 const { validationResult } = require("express-validator/check"); 
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file'); 
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -57,7 +58,6 @@ exports.postAddProducts = (req, res, next) => {
   const imageUrl = image.path; 
   // Map data to our product scheme 
   const product = new Product({
-    _id: new mongoose.Types.ObjectId("5c229d8e24e91e748a747818"),
     title: title,
     price: price,
     description: description,
@@ -144,6 +144,7 @@ exports.postEditProduct = (req, res, next) => {
     product.price = updatedPrice; 
     product.description = updatedDesc; 
     if (image) {
+      fileHelper.deleteFile(product.imageUrl); 
       product.imageUrl = image.path;
     }
     return product.save() // update object 
@@ -182,14 +183,21 @@ exports.getProducts = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
   // Getting Id from request 
   const prodId = req.body.productId; 
-  Product.deleteOne({_id: prodId, userId: req.user._id})
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found')); 
+      }
+      fileHelper.deleteFile(product.imageUrl); 
+      return Product.deleteOne({ _id: prodId, userId: req.user._id }); 
+    })
     .then(() => {
-      console.log("ITEM REMOVED FROM DATABASE"); 
-      res.redirect('/admin/products'); 
+      console.log("ITEM REMOVED FROM DATABASE");
+      res.redirect('/admin/products');
     })
     .catch(err => {
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(error); 
-    })
+      return next(error);
+    }); 
 }
